Deform

Compute the deformation by applying both bending and torsion

bend(
  data,
  elastic_modulus,
  shear_modulus,
  step = 0.02,
  points = 100,
  iterations = 15,
  verbose = TRUE
)

Arguments

data
Point data.frame (see details).

elastic_modulus
Elasticity modulus (bending, MPa)

shear_modulus
Shear modulus (torsion, MPa)

step
Length of the segments that discretize the object (m).

points
Number of points used in the grid discretizing the section.

iterations
Number of iterations to compute the torsion and bending

verbose
Boolean. Return information during procress?

Value

A data.frame():

x (m): X coordinate

y (m): Y coordinate

z (m): Z coordinate

length (m): length of the section

angle_xy (degree): angle of the segment with the XY plane

angle_xz (degree): angle of the segment with the XZ plane

torsion (degree): rotation angle of the segment

Details

The data argument is a data.frame() with each row being a point and each column being:

x: x coordinate of the segment

y: y coordinate of the segment

z: z coordinate of the segment

type: section type of the segment. 1: triangle (bottom-oriented); 2 : rectangle; 3 : triangle (top-oriented); 4 : ellipsis; 5 : circle.

width (m): segment section width

height (m): segment section height

mass (kg): mass of the segment

mass_right (kg): mass carried by the segment, on the right side

mass_left (kg): mass carried by the segment, on the left side

torsion (degree): initial torsion angle of the segment

distance_application (m): application distances for the left and right weights (see e.g. distance_weight_sine())

Examples

filepath = system.file("extdata/6_EW01.22_17_kanan.txt", package = "biomech")
df = read_mat(filepath)
# Un-bending the field measurements:
df = unbend(df)

# Adding the distance of application of the left and right weight:
df$distance_application = distance_weight_sine(df$x)

# (Re-)computing the deformation:
df_bend = bend(df, elastic_modulus = 2000, shear_modulus = 400, step = 0.02, points = 100,
               iterations = 15, verbose = TRUE)

#>  Final torsion angle at the tip (degree) =  12.52822