Optimize bending parameters

Optimize the elastic_modulus, the shear_modulus or both using observations.

optimize_bend(
  field_data,
  elastic_modulus = c(1, 100, 10000),
  shear_modulus = c(1, 1000, 10000),
  type = c("all", "bending", "torsion"),
  tol = .Machine$double.eps^0.25,
  ...
)

Arguments

field_data
Field data (data.frame of x, y and z coordinates of the points)

elastic_modulus
Elasticity modulus (bending, MPa) vector: c(min,start,max)

shear_modulus
Shear modulus (torsion, MPa) vector: c(min,start,max)

type
Type of optimization (either "bending","torsion" or "all")

tol
Tolerance for optimization accuracy

...
Further parameters to pass to bend()

Value

The optimized values

Details

The minimim and maximum values are not used when a parameter is not optimized, but they still should be provided by the user though.

Examples

file_path = system.file("extdata/6_EW01.22_17_kanan.txt", package = "biomech")
field_data = read_mat(file_path)
# optimize_bend(field_data, type = "all")
# Gives:
# $elastic_modulus
# [1] 1209.396
# $shear_modulus
# [1] 67.44096